build: Add Navigation Bar (#39)
# Pull Request

## Description

This pull request includes changes to the `GitHubPulse` project to
update the main view and improve the user interface. The most important
changes include renaming the main view and adding a list of RSS feed
items.

Updates to the main view:

*
[`GitHubPulse/ContentView.swift`](diffhunk://#diff-bb6a2edac1c9e2e3fb6fb43b49ad1ce776d7d73dc4d3eac2bc975a31caacfdfaL10-R24):
Renamed `ContentView` to `RSSFeedView`, added a state variable
`feedItems` to hold a list of RSS feed items, and updated the body to
display these items in a `NavigationView`.
*
[`GitHubPulse/GitHubPulseApp.swift`](diffhunk://#diff-b9dcef24c31c2354f24cde7b1d8f108c4f99e0bcef4500ce400b8ad292cced4dL14-R14):
Updated the main app entry point to use `RSSFeedView` instead of
`ContentView`.

Fixes #36