feat: Display RSS Titles in Navigation Bar (#40)
# Pull Request

## Description

This pull request includes significant changes to the `GitHubPulse`
project, introducing a new RSS feed fetching feature and updating the
app's entitlements for network access. The most important changes
include the addition of an `RSSFeedViewModel` for managing RSS feed
data, the creation of an `RSSParserDelegate` for parsing the RSS feed,
and updates to the `RSSFeedView` and `GitHubPulseApp` to integrate the
new view model.

### New RSS Feed Fetching Feature:

*
[`GitHubPulse/ContentView.swift`](diffhunk://#diff-bb6a2edac1c9e2e3fb6fb43b49ad1ce776d7d73dc4d3eac2bc975a31caacfdfaR9-R89):
Added `RSSFeedViewModel` class to fetch and manage RSS feed data, and
`RSSParserDelegate` class to parse the RSS feed. Updated `RSSFeedView`
to use `RSSFeedViewModel` and fetch feed items on appear.

### App Entitlements Update:

*
[`GitHubPulse/GitHubPulse.entitlements`](diffhunk://#diff-15d56784fe93fd99b0c7c98214c477f1e479e090dc6b108b9d474fb9c6e24802R9-R10):
Added network client entitlement to allow the app to fetch data from the
network.

### Integration of View Model:

*
[`GitHubPulse/GitHubPulseApp.swift`](diffhunk://#diff-b9dcef24c31c2354f24cde7b1d8f108c4f99e0bcef4500ce400b8ad292cced4dL14-R14):
Updated `RSSFeedView` initialization to pass an instance of
`RSSFeedViewModel`.

Fixes #36