ci: Add Code Quality Checks (#41)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow to automate
code quality checks. The workflow is designed to run on specific events
and includes a job to lint the codebase using the Super-Linter tool.

Key changes:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R1-R39):
Added a new workflow configuration named "Code Checks" that triggers on
push events to the `main` branch and pull request events of type
`opened` and `synchronize`. The workflow includes permissions settings
and a job named `check-code-quality` that uses the Super-Linter tool to
lint the codebase.

Fixes #11