ci: Add Markdown Link Checker (#43)
# Pull Request

## Description

This pull request introduces a new workflow to check for broken links in
Markdown files and configures the necessary settings for the link
checker tool. The most important changes include adding a new job to the
GitHub Actions workflow and configuring the link checker tool.

Configuration and workflow updates:

*
[`.github/other-configurations/.linkspector.yml`](diffhunk://#diff-f9691f23ea6c4c34ecbaa23a98721e0aa3c7a94f2fcdd38f7b99fe863c5f6305R1-R7):
Added configuration for the link checker tool, specifying directories to
include, files to exclude, and valid status codes.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R40-R56):
Added a new job named `check-markdown-links` to the GitHub Actions
workflow to check for broken links in Markdown files using the
`UmbrellaDocs/action-linkspector` action.

Fixes #13