chore: Upgrade CodeQL Versions (#60)
# Pull Request

## Description

This pull request includes updates to the
`.github/workflows/code-checks.yml` file to ensure the latest versions
of dependencies are used. The changes include updating the versions of
`setup-uv` and various `github/codeql-action` components.

Dependency updates:

* Updated `astral-sh/setup-uv` from version `v5.2.1` to `v5.2.2` to
ensure the latest version of `uv` is installed.
* Updated `github/codeql-action/upload-sarif` from version `v3.28.1` to
`v3.28.9` to use the latest version for uploading SARIF files.
* Updated `github/codeql-action/init` from version `v3.28.1` to
`v3.28.9` to initialize CodeQL with the latest version.
* Updated `github/codeql-action/analyze` from version `v3.28.1` to
`v3.28.9` to perform CodeQL analysis with the latest version.