chore: Check Justfile Format (#70)
# Pull Request

## Description

This pull request to the `.github/workflows/code-checks.yml` file
includes the addition of a new job to check the format of the `Justfile`
and updates to the versions of various GitHub Actions used in the
workflow.

New job addition:

* Added a new job `check-justfile-format` to ensure the `Justfile` is
correctly formatted. This job checks out the repository, sets up Just,
and runs the `just format-check` command.

Version updates:

* Updated the `setup-uv` action from version `v5.2.2` to `v5.3.0` in the
`run-zizmor` job.
* Updated the `upload-sarif` action from version `v3.28.9` to `v3.28.10`
in the `run-zizmor` job.
* Updated the `init` action from version `v3.28.9` to `v3.28.10` in the
`CodeQL` job.
* Updated the `analyze` action from version `v3.28.9` to `v3.28.10` in
the `CodeQL` job.