chore: Update Dependabot Groups and Configuration (#93)
# Pull Request

## Description

This pull request updates the `.github/dependabot.yml` file to enhance
the configuration for Dependabot. The changes include adding a timezone
for the update schedule, refining the grouping of updates, and excluding
specific patterns.

Enhancements to Dependabot configuration:

* Added `timezone: "Europe/London"` to the update schedule to specify
the time zone for cron-based updates.
* In the `github-actions` group, added `applies-to: "version-updates"`
to specify the scope of the group.
* Introduced `exclude-patterns` under the `github-actions` group to
exclude updates matching the pattern `"super-linter/*"`.