chore: Sunset the project (#106)
# Pull Request

## Description


This pull request updates the `README.md` file to indicate that the
project has been sunsetted and is no longer maintained. It also provides
a link to the new iteration of the project, `SourceScan`.

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R1-R5):
Added a warning notice to inform users that the project has been
sunsetted and redirected them to the new project, `SourceScan`.